brochure: Multipage 'Shiny' Apps

Build natively multi-page 'Shiny' applications, serving content on several endpoints instead of a single page. Pages are declared with their own user interface and server function and are dispatched by URL with 'routr', including parameterised routes. Redirections, request and response middleware, and helpers to read and set cookies are also provided.

See also

Author

Maintainer: Colin Fay contact@colinfay.me (ORCID) [copyright holder]

Authors:

Colin Fay contact@colinfay.me (ORCID) [copyright holder]